China-Real-Estate-Cycle-Analyzer/requirements.txt
# China Real Estate Cycle Analyzer - Dependencies

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Data & API
requests>=2.31.0
yfinance>=0.2.35

# Database (if needed)
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
